Revenue Tab fixes

+++ dashboard-services/src/main/java/org/egov/dss/repository/rowmapper/TenantWiseConnectionsRowMapper.java
@@ -8,16 +8,16 @@ import java.util.HashMap;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.ResultSetExtractor;
 
-public class TenantWiseConnectionsRowMapper implements ResultSetExtractor<HashMap<String, Long>>{
+public class TenantWiseConnectionsRowMapper implements ResultSetExtractor<HashMap<String, BigDecimal>>{
 	
 	@Override
-	public HashMap<String, Long> extractData(ResultSet rs) throws SQLException, DataAccessException {
+	public HashMap<String, BigDecimal> extractData(ResultSet rs) throws SQLException, DataAccessException {
 		
-		HashMap<String, Long> responseMap = new HashMap<>();
+		HashMap<String, BigDecimal> responseMap = new HashMap<>();
 		
 		while(rs.next()) {
 			String tenantid = rs.getString("tenantid");
-			Long totalCollection = rs.getLong("connections");
+			BigDecimal totalCollection = rs.getBigDecimal("connections");
 			responseMap.put(tenantid, totalCollection);
 		}
 		return responseMap;
